Move distribution algorithm into power distributing actor package
The power distributing algorithm is a lower level functionality, that
is used only by the power distributing actor.  So it shouldn't be part
of the public interface.
