+++ src/frequenz/sdk/actor/power_distributing/_distribution_algorithm/_distribution_algorithm.py
@@ -10,7 +10,7 @@ from typing import Dict, List, NamedTuple, Tuple
 
 from frequenz.sdk._internal._math import is_close_to_zero
 
-from ..microgrid.component import BatteryData, InverterData
+from ....microgrid.component import BatteryData, InverterData
 
 _logger = logging.getLogger(__name__)
