+++ tests/actor/power_distributing/__init__.py
@@ -1,4 +1,4 @@
 # License: MIT
 # Copyright © 2022 Frequenz Energy-as-a-Service GmbH
 
-"""Test power distribution module."""
+"""Tests for the power distributing actor and algorithm."""
